+++ packages/vite-node-miniflare/package.json
@@ -29,20 +29,23 @@
     "dev": "node ./demo/dev.mjs",
     "release": "pnpm publish --no-git-checks --access public"
   },
-  "devDependencies": {
+  "dependencies": {
     "@hattip/adapter-node": "^0.0.34",
     "@hattip/compose": "^0.0.34",
+    "miniflare": "^3.20231030.3",
+    "vite-node": "^1.0.4"
+  },
+  "devDependencies": {
     "@hiogawa/tiny-react": "0.0.2-pre.3",
     "@hiogawa/tiny-rpc": "^0.2.3-pre.11",
     "@hiogawa/utils": "1.6.1-pre.9",
     "esbuild": "^0.18.6",
-    "miniflare": "^3.20231030.3",
     "pathe": "^1.1.1",
     "vite": "^5.0.4",
-    "vite-node": "^1.0.4",
     "wrangler": "^3.19.0"
   },
   "peerDependencies": {
+    "miniflare": "*",
     "vite": "*"
   }
 }
