+++ packages/vite-node-miniflare/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./worker-script";
 export * from "./server";
+export * from "./plugin";
